requirements.txt
openai-whisper
torch
ollama
torchaudio
speechbrain
scikit-learn
numpy
soundfile 